-g16 / -g0 variants, more stripes

+++ derecho/mpiio_lustre_perf/test_stripes.sh
@@ -14,14 +14,14 @@ casedir="${topdir}/logs/default_PFL/"
 echo "submitting ${casedir}"
 mkdir -p ${casedir} && cd ${casedir} || exit 1
 
-ln -s ${topdir}/config_env.sh .
-ln -s ${topdir}/${test} .
+[ -f config_env.sh ] || ln -s ${topdir}/config_env.sh .
+[ -f ${test}       ] || ln -s ${topdir}/${test} .
 
 LASTID=$(qsub -A SCSG0001 -N ${test}-default_PFL ${topdir}/run.pbs)
 
 #--------------------------------------------
 # sweep over stripes
-for size in 128M 064M 032M 016M 004M 001M; do
+for size in 128M 064M 032M 016M 008M 004M 002M 001M; do
     for count in 02 04 06 12 24 48 96; do
         casedir="${topdir}/logs/c$(printf "%02d" ${count})_S${size}/"
         echo "submitting ${casedir}"
@@ -29,8 +29,8 @@ for size in 128M 064M 032M 016M 004M 001M; do
         stripe_cmd="lfs setstripe -c ${count} -S ${size} ."
         eval "${stripe_cmd}"
 
-        ln -s ${topdir}/config_env.sh .
-        ln -s ${topdir}/${test} .
+        [ -f config_env.sh ] || ln -s ${topdir}/config_env.sh .
+        [ -f ${test}       ] || ln -s ${topdir}/${test} .
 
         LASTID=$(qsub -A SCSG0001 -N ${test}-c${count}_S${size} -W depend=afterany:${LASTID} ${topdir}/run.pbs)
     done
